+++ plugin/nodes/network.py
@@ -2,44 +2,7 @@ from cloudify import ctx
 from cloudify.state import ctx_parameters as inputs
 from cloudify.decorators import operation
 from cloudify.exceptions import *
-
-def reset_log_indentation():
-    ctx.instance.runtime_properties['indent'] = 0
-
-def get_log_indentation():
-    if 'indent' in ctx.instance.runtime_properties:
-        return '\t'*ctx.instance.runtime_properties['indent']
-    else:
-        return ''
-
-def increase_log_indentation():
-    ctx.instance.runtime_properties['indent'] += 1
-
-def decrease_log_indentation():
-    ctx.instance.runtime_properties['indent'] -= 1
-
-def get_child(dictionary, key, required=False, debug=False):
-    child = None
-    if key not in dictionary:
-        msg = '{0} Required key "{1}" not defined!'.format(get_log_indentation(), str(key))
-        if required:
-            raise NonRecoverableError(msg)
-        else:
-            ctx.logger.debug(msg)
-    else:
-        ctx.logger.debug('{0} Obtaining key "{1}"'.format(get_log_indentation(), str(key)))
-        child = dictionary[key]
-    return child
-
-def create_child(dictionary, key, value):
-    child = get_child(dictionary=dictionary, key=key, required=False)
-    if child is None:
-        msg = '{0} New key "{1}" defined!'.format(get_log_indentation(), str(key))
-        ctx.logger.debug(msg)
-        dictionary[key] = value
-        child = dictionary[key]
-    return child
-
+from plugin.nodes.utils import *
 
 def build_radl_network(config):
     ctx.logger.debug('{0} Infrastructure Manager deployment info:'.format(get_log_indentation()))
