+++ plugin/relationships/utils.py
@@ -1,0 +1,39 @@
+from cloudify import ctx
+from cloudify.exceptions import *
+
+def reset_log_indentation():
+    ctx.source.instance.runtime_properties['indent'] = 0
+
+def get_log_indentation():
+    if 'indent' in ctx.source.instance.runtime_properties:
+        return '\t'*ctx.source.instance.runtime_properties['indent']
+    else:
+        return ''
+
+def increase_log_indentation():
+    ctx.source.instance.runtime_properties['indent'] += 1
+
+def decrease_log_indentation():
+    ctx.source.instance.runtime_properties['indent'] -= 1
+
+def get_child(dictionary, key, required=False, debug=False):
+    child = None
+    if key not in dictionary:
+        msg = '{0} Required key "{1}" not defined!'.format(get_log_indentation(), str(key))
+        if required:
+            raise NonRecoverableError(msg)
+        else:
+            ctx.logger.debug(msg)
+    else:
+        ctx.logger.debug('{0} Obtaining key "{1}"'.format(get_log_indentation(), str(key)))
+        child = dictionary[key]
+    return child
+
+def create_child(dictionary, key, value):
+    child = get_child(dictionary=dictionary, key=key, required=False)
+    if child is None:
+        msg = '{0} New key "{1}" defined!'.format(get_log_indentation(), str(key))
+        ctx.logger.debug(msg)
+        dictionary[key] = value
+        child = dictionary[key]
+    return child
